+++ vscode/CHANGELOG.md
@@ -1,5 +1,22 @@
+---
+title: Changelog
+layout: page
+description: VS Code changelog
+parent: VS Code
+grand_parent: Plugins
+permalink: /plugins/:path/changelog
+---
+
 # Changelog
 All notable changes to VS Code plugin will be documented here.
+
+## 4.14.1-1 - 2023-07-26
+### Added
+- Preinstallation of platformio-ide and thinger-io extension
+- Default settings for Thinger.io extension configuration
+
+### Changed
+- Bumped to code-server version 4.14.1
 
 ## 4.11.0-3 - 2023-03-31
 - Updated deprecated mod for code-server
@@ -13,3 +30,4 @@ All notable changes to VS Code plugin will be documented here.
 ## 4.6.0-1 - 2022-08-26
 - Downsized resource memory to 3.75GB
 - Bumped to code-server version 4.6.0
+